Pick commit from global-template
This corresponds to commit
7fc89ed0841cba2e6d554302a74a2e38686e248f
from the global-template branch.

If you want to update this repository with the latest commit from
global-template, you have to create a special commit before merging it
into this repository.
To do this, create a commit whose parent is this commit, and whose state
corresponds with the tree of the latest commit in global-template. Then
you can merge that new commit into this repository, as if merging
global-template instead.